Some changes after review and code simplification
